+++ planning/planning_pipeline/src/planning_pipeline.cpp
@@ -214,7 +214,7 @@ bool planning_pipeline::PlanningPipeline::generatePlan(const planning_scene::Pla
   if (publish_received_requests_)
     received_request_publisher_.publish(req);
   adapter_added_state_index.clear();
-
+  
   if (!planner_instance_)
   {
     ROS_ERROR("No planning plugin loaded. Cannot plan.");
@@ -232,7 +232,7 @@ bool planning_pipeline::PlanningPipeline::generatePlan(const planning_scene::Pla
         std::stringstream ss;
         for (std::size_t i = 0 ; i < adapter_added_state_index.size() ; ++i)
           ss << adapter_added_state_index[i] << " ";
-        ROS_DEBUG("Planning adapters have added states at index positions: [ %s]", ss.str().c_str());
+        ROS_INFO("Planning adapters have added states at index positions: [ %s]", ss.str().c_str());
       }
     }
     else
@@ -299,8 +299,8 @@ bool planning_pipeline::PlanningPipeline::generatePlan(const planning_scene::Pla
             {
               // check validity with verbose on
               const robot_state::RobotState &kstate = res.trajectory_->getWayPoint(index[i]);
-              planning_scene->isStateValid(kstate, req.group_name, true);
-
+              planning_scene->isStateValid(kstate, req.path_constraints, req.group_name, true);
+              
               // compute the contacts if any
               collision_detection::CollisionRequest c_req;
               collision_detection::CollisionResult c_res;