+++ planning_interface/move_group_interface/src/move_group.cpp
@@ -408,6 +408,22 @@ public:
     return active_target_;
   }
 
+  bool startStateMonitor(double wait)
+  {
+    if (!current_state_monitor_)
+    {
+      ROS_ERROR("Unable to monitor current robot state");
+      return false;
+    }
+
+    // if needed, start the monitor and wait up to 1 second for a full robot state
+    if (!current_state_monitor_->isActive())
+      current_state_monitor_->startStateMonitor();
+
+    current_state_monitor_->waitForCurrentState(opt_.group_name_, wait);
+    return true;
+  }
+  
   bool getCurrentState(robot_state::RobotStatePtr &current_state, double wait_seconds = 1.0)
   {
     if (!current_state_monitor_)
@@ -1465,6 +1481,11 @@ void moveit::planning_interface::MoveGroup::rememberJointValues(const std::strin
   rememberJointValues(name, getCurrentJointValues());
 }
 
+bool moveit::planning_interface::MoveGroup::startStateMonitor(double wait)
+{
+  return impl_->startStateMonitor(wait);
+}
+
 std::vector<double> moveit::planning_interface::MoveGroup::getCurrentJointValues()
 {
   robot_state::RobotStatePtr current_state;
@@ -1477,7 +1498,7 @@ std::vector<double> moveit::planning_interface::MoveGroup::getCurrentJointValues
 std::vector<double> moveit::planning_interface::MoveGroup::getRandomJointValues()
 {
   std::vector<double> r;
-  impl_->getJointModelGroup()->getVariableRandomValues(impl_->getJointStateTarget().getRandomNumberGenerator(), r);
+  impl_->getJointModelGroup()->getVariableRandomPositions(impl_->getJointStateTarget().getRandomNumberGenerator(), r);
   return r;
 }