+++ planning_interface/CHANGELOG.rst
@@ -2,6 +2,10 @@
 Changelog for package moveit_ros_planning_interface
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+0.5.8 (2013-10-11)
+------------------
+* add function to start state monitor in move_group_interface::MoveGroup
+
 0.5.7 (2013-10-01)
 ------------------